+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-beta.108
+
+### Patch Changes
+
+- Updated dependencies [c7c3296]
+  - sit-onyx@1.0.0-beta.108
+
+## 1.0.0-beta.107
+
+### Patch Changes
+
+- Updated dependencies [357ac46]
+  - sit-onyx@1.0.0-beta.107
+
+## 1.0.0-beta.106
+
+### Patch Changes
+
+- Updated dependencies [9e53a89]
+  - sit-onyx@1.0.0-beta.106
+
 ## 1.0.0-beta.105
 
 ### Patch Changes