+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,41 @@
 # sit-onyx
 
+## 1.0.0-beta.108
+
+### Minor Changes
+
+- c7c3296: feat: Expose input elements of form elements
+
+## 1.0.0-beta.107
+
+### Major Changes
+
+- 357ac46: feat(OnyxDataGrid): implement menu items for sorting feature
+
+  #### Breaking changes
+
+  - OnyxDataGrid: rename header actions property `listItems` to `menuItems`. It now expects `OnyxMenuItem` components instead of `OnyxListItem`
+
+  #### Features
+
+  - OnyxDataGrid: add German translations
+
+  #### Fixes
+
+  - OnyxDataGrid: update translations when locale changes
+  - OnyxUserMenu: use `OnyxMenuItem` for footer instead of `OnyxListItem`
+
+## 1.0.0-beta.106
+
+### Major Changes
+
+- 9e53a89: feat(OnyxStepper): update OnyxStepper API
+
+  - remove property `step`, use `stepSize` instead
+  - remove property `stripStep`, use `validStepSize` instead. User inputs will no longer be manipulated, instead an error will be shown
+  - changed logic of `precision` property. Now determined numbers of decimal places to show. Is no longer the default value of `stepSize` property.
+  - fix bug that decimal value is not displayed correctly when `precision` is not set
+
 ## 1.0.0-beta.105
 
 ### Patch Changes